requirements.txt
asn1crypto==0.24.0
astroid==1.6.2
bottle==0.12.13
Cerberus==1.1
cffi==1.11.5
gunicorn==19.7.1
idna==2.6
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pycparser==2.18
PyJWT==1.6.1
pylint==1.8.3
pymongo==3.6.1
python-dotenv==0.8.2
six==1.11.0
wrapt==1.10.11
